Move EventDispatcher to base dispatch package
